+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - removed newClickHandler, now relying only on mousedown/mouseup event (simpler touch support)
 - added editable:drawing:move
 - added drawingCursor option to L.Editable
+- added editable:vertex:metakeyclick event
 
 ## 0.5.0
 - added editable:vertex:drag event